[add] GPUでもAndroidが通るように実装を追加
