+++ .github/workflows/build-gpu-android-lib.yml
@@ -1,0 +1,110 @@
+name: 'GPUのビルド(Android)'
+
+on:
+  workflow_dispatch:
+  push:
+    paths:
+      - 'buffer/gpu/rs/**'
+      - '.github/workflows/build-gpu-android-lib.yml'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    name: Build (Android)
+
+    steps:
+      - name: チェックアウト
+        uses: actions/checkout@v4
+
+      - name: Javaのセットアップ
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'zulu'
+          java-version: 17
+          cache: 'gradle'
+
+      - name: gradlewに権限を付与
+        run: chmod 777 ./gradlew
+
+      - name: キャッシュを復元
+        uses: actions/cache@v4
+        with:
+          path: buffer/gpu/rs/target
+          key: gpu-android-${{ hashFiles('buffer/gpu/rs/Cargo.toml') }}
+
+      - name: Android NDK のセットアップ
+        uses: nttld/setup-ndk@v1
+        with:
+          ndk-version: r28c
+          add-to-path: true
+
+      - name: Rust の Android ターゲットを追加
+        run: rustup target add aarch64-linux-android x86_64-linux-android
+
+      - name: NDK クロスコンパイラを PATH に追加
+        run: echo "$ANDROID_NDK_ROOT/toolchains/llvm/prebuilt/linux-x86_64/bin" >> $GITHUB_PATH
+
+      - name: ネイティブライブラリをビルド (arm64-v8a)
+        run: ./gradlew buffer:gpu:rs:cargoAndroidArm64Build
+
+      - name: ネイティブライブラリをビルド (x86_64)
+        run: ./gradlew buffer:gpu:rs:cargoAndroidX64Build
+
+      - name: 成果物をアップロード (arm64-v8a)
+        uses: actions/upload-artifact@v4
+        with:
+          name: gpu-android-arm64-v8a
+          path: buffer/gpu/rs/target/aarch64-linux-android/release/libgpu.so
+          if-no-files-found: error
+
+      - name: 成果物をアップロード (x86_64)
+        uses: actions/upload-artifact@v4
+        with:
+          name: gpu-android-x86_64
+          path: buffer/gpu/rs/target/x86_64-linux-android/release/libgpu.so
+          if-no-files-found: error
+
+  commit:
+    needs: build
+    runs-on: ubuntu-latest
+    name: Commit native libraries
+
+    steps:
+      - name: チェックアウト
+        uses: actions/checkout@v4
+
+      - name: 全成果物をダウンロード
+        uses: actions/download-artifact@v4
+        with:
+          path: artifacts
+
+      - name: アーティファクト構造を確認
+        run: |
+          echo "=== Artifacts structure ==="
+          find artifacts -type f
+
+      - name: 成果物を jniLibs に配置
+        run: |
+          # arm64-v8a
+          mkdir -p buffer/gpu/src/androidMain/jniLibs/arm64-v8a
+          find artifacts/gpu-android-arm64-v8a -name "libgpu.so" -exec cp {} buffer/gpu/src/androidMain/jniLibs/arm64-v8a/ \;
+
+          # x86_64
+          mkdir -p buffer/gpu/src/androidMain/jniLibs/x86_64
+          find artifacts/gpu-android-x86_64 -name "libgpu.so" -exec cp {} buffer/gpu/src/androidMain/jniLibs/x86_64/ \;
+
+          echo "=== jniLibs structure ==="
+          find buffer/gpu/src/androidMain/jniLibs -type f
+
+      - name: 変更をコミット
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          git pull
+          git add buffer/gpu/src/androidMain/jniLibs/
+          if git diff --staged --quiet; then
+            echo "No changes to commit"
+          else
+            git commit -m "[CI] Update GPU(Android) libraries"
+            git push
+          fi
