+++ buffer/gpu/rs/.cargo/config.toml
@@ -1,0 +1,5 @@
+[target.aarch64-linux-android]
+linker = "aarch64-linux-android28-clang"
+
+[target.x86_64-linux-android]
+linker = "x86_64-linux-android28-clang"
