+++ buffer/gpu/src/androidMain/kotlin/com/wsr/gpu/GPUBackend.android.kt
@@ -1,0 +1,8 @@
+package com.wsr.gpu
+
+import com.wsr.base.IBackend
+
+actual fun loadGPUBackend(): IBackend? = runCatching<IBackend> {
+    System.loadLibrary("gpu")
+    GPUBackend()
+}.getOrNull()
