+++ buffer/gpu/src/jvmAndroidMain/kotlin/com/wsr/gpu/elementwise/math/JMath.kt
